Correct fix for zen_get_customid to address individual entries.
Corrected a correction for zen_get_customid.  Previously an error was generated by passing an array into the sql for pulling all of the customids of a product that was built using individual option name/option value combinations.  The "solution" proposed did not provide the individual customids to be pieced together to give a single string.  This modification rectifies that and streamlines the overall process.