Revised the admin file admin/products_with_attributes_stock.php to support passing security scans.
One issue that stands out from this code as compared
to ZC admin files that begin/have a $_GET['action']
section is that there is a zen_not_null operation
that encapsulates the $action section.  Even though
there was appropriate testing and setting, this is not
recognized by some scan software.  Therefore, this
file was modified to agree with standard ZC file
formatting.