Remove extraneous zen_get_products_stock code
In includes/functions/functions_lookups.php there was extraneous code that was "functional" but made the zen_get_products_stock function complicated.  An improved function had been implemented and circumvented the need for the additional code, but that additional code was desired not to be lost in the shuffle if problems arose.

That code is identified as no longer necessary and removed by this commit.