Incorporate function into zen_get_products_stock
For some reason when putting together the package for ZC 1.5.5, a single return function was omitted from the package when determining the available stock of a product.  This function had been written to address/handle returning stock quantities related to other attributes and their associated designators so that comparison could be made against in cart designators.  Anyways, this function is incorporated back into the code more than likely making the new function referenced one line below the inserted function return in the extra_functions folder unnecessary.