Strip first file upload review
Removed code that would otherwise store the file(s), though at the moment the goal is to simply validate that the provided attributes are sufficient/covered by SBA. Because need to also include the uploaded file attribute(s) in the review.  The alternative is to not populate SBA with file upload attributes and allow just those that fall under the 'id' type to "prevail".  This later part would require modifying the SBA table admin population routine.