Add the class products_with_attributes_stock as an admin variable
In some cases in the admin, a variable is initiated/reset to access the class data of the class products_with_attributes_stock.  This will initiate the variable and make it available as a global variable to be used as necessary rather than having to include the code to initiate it.