Correct usage of RecordCount
In the function: insertSBAproductsOptionsTypes the check of the count of records associated with the basic SBA option name type already existing incorrectly used RecordCount instead of the function name RecordCount(). This should have only affected stores running the installation a second time.

Also, as the test for action in that section is if a database configuration constant is defined or not, once the constant has been successfully added to the database it is defined in the code so that future evolutions of the install do not try to perform the work again.