define the correct value in process of installation.
Apparently was looking at the wrong information when trying to populate the define for PRODUCTS_OPTIONS_TYPE_SELECT_SBA and did not assign the value that is expected to be used.  This would have been a "temporary" assignment, but was wrong regardless.  It is now assigned the numerical value that ought to be used into the future after installation.