Modify removal of SBA to adjust option names to use default dropdown
Installation of SBA adds an option name data type entry selection of SBA Select List (Dropdown) Basic; however, because removal of the plugin is supposed to also remove entered SQL data, product that have had this option selected need to be revised to an alternate method when the plugin is removed.  Because this is a modified/upgraded version of the standard dropdown, restoring the option back to a standard dropdown seems appropriate and should be performed by this yet untested commit.