Hide DD attributes marked as Display Only, Default and Sort = 0
Added code/information for pad_sba_sequenced_dropdowns
so that if an attribute is marked as display only, Default
and sort order of 0 will not be displayed in the
generated Dynamic Dropdown menu.  This is a typical
"feature" of attributes that attempt to direct
the customer to select an item; however, DD already
performs this action automatically.