Reuse $products_with_attributes_stock_class variable
Instead of declaring a new variable as the class
developed to support SBA, the class as set at the
onset of loading the store is used and the various
locations updated where this was used.  Primarily
this is a memory use/access improvement as the
clas does not need to be declared/defined just
further accessed as a global variable.