Clean up code of insertSBAproductsOptionsTypes
Reports have been made that the function insertSBAproductsOptionsTypes is not adding the SBA Select List (Dropdown) Basic option to the option names area of the admin.  While the code appears to provide the desired result, the results have been indicating otherwise.  Continuing to try to automatically add this feature to support SBA without having to run one or more external SQL queries which have shown to work.  This implies that there is something wrong with the code.  In this iteration have cleared the cache for when looking for and modifying the products options types and have also removed other atypical ZC code formatting although historically that coding has not caused a problem which is why it was never before modified.
Also added additional query control with the use of bindVars and a variable to be used throughout for reference to the dropdown name.  This dropdown name probably should/could be moved to a language define so that this feature can be better understood in other languages.