Update zen_register_admin_page for < ZC v1.5.4
In incorporation of the ZC admin function zen_register_admin_page, consideration was not provided for ZC 1.5.4 and below where the sort order must be passed to the function.  In ZC 1.5.5, the function auto-determines the next sort_order if one is not provided and this was how this code was originally written when incorporating zen_register_admin_page.

An error is thrown in ZC 1.5.4 without the sort order being provided, which is now corrected/made backwards compatible by this commit.