Multi-attribute stock not decreasing after purchase
The function zen_get_sba_stock_attribute was not identifying the attribute array when processing data from an order (used/accessed on checkout) and this resulted in the stock level not decreasing when a multiple attribute variant product was purchased.  Single attribute stock would decrement and one of the stock quantities would go down if all attributes were separate, but not stock for all attributes would go down. Currently this resolves the issue for a multiple attribute variant.  Having multiple individual attributes for a product remains to be reviewed.