Pull code from checkout_shipping header as pre-code
Attempting to remove code from core files, and the checkout_shipping header file is one of those files that has been modified to address a "bigger" issue that when checkout is progressing to validate that the cart has correct product before allowing the final purchase.  This can be accomplished a number of ways, the preferred way for more recent versions of ZC are to use the notifier in the applicable location; however, pre 1.5.3 this would often require additional work as the notifiers could not pass alternate/modifiable information without some code replacement or inheritance.  Trying to keep this module (for now) compatible with ZC 1.5.1, but will be abandoning that at some point in the future.