Repair 1.5.5 admin invoice operation
Apparently in manipulating variables through the code, the admin/invoice.php file for ZC 1.5.5 referenced a non-existent class that wasn't declared.  Corrected this in this commit for ZC 1.5.5.  Appears that variable had been changed to reduce the need to declare a variable that already existed; however, the wrong variable was referenced in each of the invoice files.

Partially Fixes #31 for ZC 1.5.5. Need to address for other ZC versions.