Repair ZC 1.5.1 admin/invoice.php
Apparently in manipulating variables through the code, the admin/invoice.php file for ZC 1.5.1 referenced a non-existent class that wasn't declared.  Corrected this in this commit for ZC 1.5.1.  Appears that variable had been changed to reduce the need to declare a variable that already existed; however, the wrong variable was referenced in each of the invoice files.

Fixes #31 for ZC 1.5.1 and completes all fixes. Addressed in all other ZC versions before this commit.