Update class.products_with_attributes_stock.php
Apparently in manipulating variables through the code, the admin/includes/classes/observers/class.products_with_attributes_stock.php file for all ZC versions referenced a non-existent class that wasn't declared.  Corrected this in this commit.  Appears that variable had been changed to reduce the need to declare a variable that already existed; however, the wrong variable was referenced in the usage of the class.  The class had been properly passed forward, but was incorrectly referenced.

Fixes Issue #32 for all ZC versions.