ZC 1.5.5: Remove modifications to includes/modules/pages/shopping_cart/header_php.php
Basically the code modifications included in the file
have been commented out to restore operation back to
the original file.  File should not be merged with
a new site, but instead issue will be addressed
via observer.  In older versions currently the info
will be processed twice, once within the module file
once in the observer.