Low Stock Email improvements
Two areas addressed.  By code review the carry over/assignment of the applicable remaining stock quantity to be addressed by this observer did not force a new assignment on each iteration/verification of the product being tracked by SBA.  In some very rare circumstance it could send an email when the quantity was not actually out of stock, but this would also require the product's information to not contain any attribute information (as a result of some programming error where the product is SBA, but the product provided to the function did not have any attributes.)
The second area addressed was the "detection" of whether previous low stock information had been gathered.  In the cycle of processing, if the total number of items is less than the limit, initial information is added, but this does not account for product that is offered in mixed quantities or in the case of SBA, it could be that a variant is low in stock, but the overall quantity of product is not.  The modification made above will expect that the variant quantity be maintained above the notification end point or else a notification will be made when that particular variant is purchased until the available quantity for the variant exceeds the ZC store limit.