Prevent duplicate qty of unchecked SBA items
This is a temporary fix; however, if a product is marked to be SBA and it only has checkbox attributes, then if the product was added to the cart with no checkboxes marked, it would be added in quantities twice that entered (once processed through the SBA extra_cart_actions file and then once through the standard ZC add-to-cart processing).

This temporary fix prevents the duplicate addition; however, is likely to be modified later to support more control over the attributes associated with the product that can be added.  This is done in the nature of improved customer experience.