Add support for READ ONLY and check box populating as well as improved stock dependency.
Read Only attributes are offered as possibly being stock dependent
though the default is for them to be non-stock dependent.  If
enabled, other plugins may not respect the stock dependency.

The big difference with that is as a stocked item, the read only
attributes appear with the product on things such as the invoice.

Conflicts:
	admin/stock_by_attr_install.php
	includes/classes/pad_sba_sequenced_dropdowns.php