Add PRODUCTS_OPTIONS_TYPE_ATTRIBUTE_GRID
Apparently Edit Orders has been expecting the product
attribute grid (PAG) plugin to generate the appropriate
constant to support that plugin and EO "self" defined
what the option type name should be. Well, time to update
the PAG plugin to work with EO. Before doing that, SBA
should be updated as well... This is just one of the files
to be modified.