Remove zc155 incontext code (#44)
* Remove unincorporated PayPal InContext Code

Partially resolves #43.

In development of the changes for ZC 1.5.5, there was an expectation that other proposed code to further support use of PayPal InContext would be incorporated.  As such, the same code was carried over to the ZC 1.5.5 version of the same modified file.  As of ZC 1.5.5d, the change has not been incorporated and is causing the error that is described at issue #43.  As such, this code is being restored to be closer to what is really the ZC 1.5.5 version of includes/modules/templates/template_default/templates/tpl_checkout_confirmation_default.php

* Remove responsive PayPal Incontext

This finalizes and fixes #43 by removing the submit button code that had been anticipated being added as part of [ZC Issue #1192](https://github.com/zencart/zencart/issues/1192).