Class cache query clear
As reported in the ZC forum for this plugin, there have been reported occurrences of stock decrement not occurring as expected.  While no specific conditions have been provided to reproduce the issue (making resolution of the issue difficult to identify), there are a few occurrences of performing a query within a loop that may have or may get cached and therefore provide incorrect final results.  This commit is an attempt to rectify that issue.