Auto-populate selections that exist in the cart when products_id supports.
It has been discovered that when a SBA tracked product includes a read-only attributes that are interspersed throughout the selections that the auto-selection of the options ends at/after the first read-only attribute.  This is supposed to restore the ZC default capability by processing the page upon first/return arrival. Ideally, this could be done within each iteration generated so that no closing execution would be needed, but there is something about that which I want to think about more.