Correct admin variant addition sorting.
When adding a combination variant of attributes, after selecting
the store/save button the posted data was sorted in a way that
could disrupt the original sort order of attributes and make
it difficult for the admin to understand the sequence of items.

This correction makes it such that all of the options are initially
sorted by the first option name and value then once all
options have been displayed, then the second attribute is
played out and so on, such that the last attribute is the one
that looks like it switches frequently.

Fixes https://github.com/potteryhouse/stock_by_attribute_1.5.4/issues/41