Populate admin order class with customid.
This adds the customid from the products with attributes stock
class/table into the order class to make the data accessible
for display, storage, etc...  It is collected by use of the
notifier/observer classes and then based on customid being
defined as a unique key, when adding customid information to
each individuall attribute, the combination of information
is treated to provide a product customid.  For example a blank
customid is not factored into a comma separated list but instead
ignored.  Thus in a list of three attributes for a product,
one could have a customid with the rest blank and the product
will be identified as having a single customid not a comma
separated list of three items with only one populated.

Fixes https://github.com/potteryhouse/stock_by_attribute_1.5.4/issues/25