Admin packingslip collects customid from historical data
Converted the use of customid information coming from active database data
to using the values that have been stored through the purchase process.

Yes that information was at one point live data; however, this means to
capture the data in the purchase history so that if it changes on the store
front, a review of past data would show the then assigned value not the
current value.

Conflicts:
	1_5_1/admin/packingslip.php