Prevent addition of display only attributes.
When determining items to be added to stock, display only
attributes are not selectable in the grand scheme of things and
therefore should not be presented to be available for adding
stock quantities.

Note: when an attribute is display only, default and sort order
of 0, then the option value will not be displayed in dynamic
dropdowns.