Checkout Confirmation compression.
Now with the customid being more dependendably returned the
functions that had been created and stuck into other template
files can now be better used from notifiers.  Also this uses the data
that is available at that moment (such as order history data) to support
display of the customid as it is what is going to be stored.