Revert to standard ZC products_id casting.
In the admin file products_with_attributes_stock.php there
remained an assignment of a products id variable that used
a function to force the type cast to an integer; however,
some scanning programs consider this as a possible issue and
prevents loading of the file to the server.  Therefore,
changed this to a direct type cast of (int) which accomplishes
the same thing, but more visibly for the inspection software.