Add single record to Non Stock table and correct spelling in configuration option.
Temporary fix to an issue was identified by ensuring at least one
record existed in the TABLE_PRODUCTS_WITH_ATTRIBUTES_STOCK_ATTRIBUTES_NON_STOCK
table, otherwise in a multiple attribute situation all option values for the
first option name tended to show as "out-of-stock" but didn't prevent
adding the product.  The issue being that the product was incorrectly
marked as out-of-stock. This relates to the logic in the dynamic dropdowns
feature/function developed for this version of SBA.