Correct array merge to account for post incrementing.
While the merging of the data was needed, the value of $i within the
observer was incorrectly considered.  $i within the observer had not
been incremented yet and therefore attempts to decrement the value
would result in non-existent arrays.  This led to odd duplication and
display on the store front for single attribute product that had some
out-of-stock attributes to be displayed.