Use array merge in place of replacement.
This ensures continued operation of any other plugin that has modified
the products_options_array so that sba doesn't just take it over.  Of course
there is an expectation that the array is being properly upheld and that
if the array is at all trimmed or elements removed that other parts
of the tracking are maintained so that operations do not get squirrely.