Modify use of empty($var) to ensure that $var also is an array.
These changes are successfully working on a server; however, haven't seen/heard
any specific errors related to not having this logic. It just seemed
unsettling that $var could be some other type of variable and the
code associated with the if would execute and/or provide errors.

So I added the additional characteristic check to be sure that things would
be handled as expected.