ALLOW_STOCK_CHECKOUT, out-of-stock marking - Addresses #50
This added a new define STOCK_MARK_ALLOW_MIX_TOTAL_ALL that needs to be
incorporated into the admin configuration screen as an SBA specific define.

This allows control of the marking of out-of-stock product in the specific
condition of:
- the stock is allowed to sell beyond the available quantity (STOCK_ALLOW_CHECKOUT === 'true'),
- the product is set to have Product Qty Min/Unit Mix set to true, AND
- the variant quantity in the cart exceeds total stock quantity of the product.

Similar control for STOCK_ALLOW_CHECKOUT == 'false' has not been incorporated yet.
And ideally could use this new define to handle all conditions rather than
adding yet another control.

When evaluating the need to modify $flagAnyOutOfStock and to modify
the value it seems that the best fit with any other plugins is to
only modify the value when it has to be cleared. This way any setting
established by another plugin (or if someone has chosen to modify it
within this plugin) will remain unless there is not to be
any product that is marked out-of-stock.