Reinstate stock reporting on shopping cart
Based on reporting and request, the quantity available of a variant
has been restored in the observer class associated with SBA product.

Part of this reporting includes identification of product that is
considered in low stock quantity with the current designation being the
quantity at which reorder notification is provided and is currently
set to the admin's define for STOCK_REORDER_LEVEL but could easily
be reassigned to a value specific to a product, variant, or otherwise
controlled.

Fixes #51