ZCv156 Reduce the amount of change needed when incorporating zen_get_products_stock
Previously, the code provided in SBA would use sanitization via bindVars
even though sanitization was provided by type-casting the variable to
an integer.  To make integration of the plugin a little easier, adding
to the core code is being done instead of modifying it.  This commit
reverses a change made a long time ago and should be applied for all
ZC versions.