Simplify the checkout_shipping header code incorporation.
Previous code offered multiple avenues for choices or direction; however,
only one has really served a need.  Because of this, the code has
been reduced down to the currently identified need to allow product
tracked by SBA to have a total quantity identified less than the quantity
requested and allow the product to move forward and to return to the
shopping cart if it is identified that the variant quantity does not exist.
There is some action taken by the notifier as well.