Update class.products_with_attributes_class_stock.php
The class function `zen_draw_pull_down_menu_SBAmod` was oversanitizing information in the string to be displayed by use of zen_output_string_protected for the entire string.  This oversanitization caused html entity strings provided as `&pound;` to be sanitized to `&amp;pound` that then when displayed would again appear as `&pound;` instead of the symbol for the British pound.

This commit restores previous functionality that had not previously been observed (reported?) as being a problem.