update products_with_attributes_stock.php remove on duplicate
In the forum a store owner identified that there was a problem with installation based on the key being too long.  The original (re)design of SBA placed a primary key as well as multiple unique keys against the PWAS table.  That uniqueness can be carried out through program design instead of relying on the database.  As such, this commit reworks the database insertion to consider factors that would provide the same level of uniqueness, but doing so via php instead of mysql.

Ultimately, it may be desired to support the table not having a unique customid; however, additional work is necessary to provide a good customer interface to transition from duplicate keys to unique keys in the event such a switch is desired.  Until then, do not want to allow non-duplicate keys for general use.  If it is strongly desired, one would need to rework the code of this commit to make it possible.