Incorporate changes to copy attribute notifier of ZC 1.5.6
The notifier adopted for ZC 1.5.6 incorporated the $pInfo variable as the first parameter.  This allowed not referencing $pInfo globally. Further the handling of the $contents variable was modified to support removing a "known" number of items off of the end of the list so that the SBA option(s) can be added and then whatever was removed added back on.