Correct sequence for copying attributes
Commented on this in the previous commit; however, the sequence of rebuilding `$contents` from removing previous data off of the top of the array put items back on top "backwards".  This was noticed in code review and would not have been identified by operation until more than one item was to be removed off of the top of the array.