Speed up processing
There are a few queries that are repeatedly run (is this product stocked
by attributes?).  This is reduces the execution of that
query and associated queries.  This has been identified to cut the time
involved in querying a product that has two option names with a minimum
of 4 each by approximately a factor of 7.