Update for sql_mode=ONLY_FULL_GROUP_BY
Reported in the Zen Cart Forum on this plugin: https://www.zen-cart.com/showthread.php?47180-Stock-by-Attribute-v4-0-addon-for-v1-3-5-1-3-9&p=1350237#post1350237 that while using MySQL 5.7 and the sql_mode was set to ONLY_FULL_GROUP_BY, the area of code in question caused a myDEBUG error.  This commit resolves the issue for this file.