Remove potential ONLY_FULL_GROUP_BY mysql error
During review of operations while sql_mode was set to ONLY_FULL_GROUP_BY, identified that the installation file had a query for the latest value used within the database, although that value is controlled by AUTO_INCREMENT and further the returned value is not being used. Removed the code within the query to prevent the error and not to ask for the information that had been sought.