Update the content storage conversion
Between the code written to collect user input and to store the associated
data, quotes were being added to the content and/or converted when
not needed.  This aligns the format of the provided data with the
intended storage format until a conversion process upstream is modified
away from its current usage.