Refactor includes/extra_cart_actions/stock_by_attributes.php
Modified sequence of data check to try to apply simple check first to
reduce processing load.
Added space between if and parentheses where one was missing.
In some cases used both the `isset` and `array_key_exists` function to
detect existence as isset can be faster.  This was not applied throughout.
Tried to reduce else options of if/else.
Reduced some logic testing such as changing `!exists || (exists && something)`
to just `!exists || something`.  if `not exists` fails, that means that it
exists and therefore not necessary to test for existence again.
Removed the direct call to `unset($fileVar['file_located'])` because
$fileVar is set to an empty array, therefore should be effectively unset.
Added a strict error processing check for customer_id.