Update handling of attribute images
In rearranging code, the variable `$options_attributes_image` became reset on initial pass of every option_name, this causes the attribute image to not be displayed for all image placement settings within the option name manager when the product is not tracked by SBA.

Removing the line from a default installed store with this plugin restored the operation of the attributes images and also allowed modifying the display of attribute images based on the setting of `SBA_SHOW_IMAGE_ON_PRODUCT_INFO`; however, all products that had attribute images were affected by the constant. Added a test for if the product is an SBA product to address the associated image controls.

This again fixes #15 which somehow became an issue again.