Correct single option name display on ZC 1.5.1 and below.
In versions before ZC 1.5.3, the observer class only recognized the
update function.  In the update function variables needed must be obtained
from the global space if they are not passed in the array or part of the
class that called the observer.  An incorrect global variable was used
in attempting to incorporate operations that affected single option
name style attributes. (If multiple option names were used, then there
were different results as there is/was a different path to display the
data.)