Restore catalog side customid for ZC 1.5.1.
With the continued development to move code out of the base software, the inability to continue testing on ZC 1.5.1 and below, and not having had much feedback related to Zen Cart 1.5.1, some functionality had been lost.  This commit is intended to restore operations associated with the customid as pertains to the observer class.  A modification is still necessary in the orders class to support some of the handling provided when a query is run on a specific order number.