Correct customid of combined attributes
In admin collection of customid information, the option_id => value_id pair was not being used when evaluating the product purchased.  As a result, `$orderClass->products[$index]['attributes'][$key]['customid'] would get incorrectly populated and as a result the processing code to display the attribute(s) would use the incorrect information.

This commit populates the above variable only if the pair was met.