Replace $sba_add_prods with array of arrays
There are a series of individual variables that are used together to handle product updates in the shopping cart.  As these are arrays, it makes sense to use a single array that has keys that can be managed instead of individual variables.

This is the first of several commits to handle the replacement of those variables.