Remove commented debug code
There was a lot of debug type code (`$_SESSION` related) that is not pertinent to execution.  Removed this excess code.

There remains commented code; however, it seems that it is in the original code base from which some of this is copied.  This way when comparing the code again, it can be identified what code should remain commented.