Address PHP notice for potential message
If `$adjust_max == 'true'` then a message is to be displayed that requires `$requested_qty`.  In the condition of the `$new_qty` being greater than `$add_max` (maximum that can be added) and the product is not marked to support mixed quantities, the value `$requested_qty` was not assigned.  This corrects that issue.