Adjust load points and add comment
Identified in ZC 1.5.6 that the `zen_products_stock` notifier was not observed when the shopping_cart was being modified. This moves the observer to a point before the load of `init_cart_actions` which is what processes the shopping_cart actions.  Also aligned the execution of the class' reset to the same load point as the class while maintaining it after the point (in array population) where the class is loaded and kept the observer after the class to ensure loading the observer does not throw notices or errors when trying to access variables that would not have otherwise been defined on first page load.

Fixes #59