Add display of attribute discount information using ZC 1.5.6 style
The code functions to display attribute prices was modified in ZC 1.5.6
to broaden the support of accurate attribute price display.  This
changed the code used to call for the attribute price.  That modification
is finally recognized here and discounted attribute prices are supported.
This includes attribute quantity discounts as well as sales, specials and
the like.