Capture code consideration for single value option name.
There was some discussion that the code did not properly handle an
option name that only had a single option value. To further modify
the system response some alternate ideas were considered which are shown
in this code section.  Use of this code caused issues with operation
where product could be added with the attribute or as a product without
any attributes depending on how the code was applied.