Adds an auto-observer (ZC 1.5.3 and above) to support searching on customid.
Adapts a copy of the advanced_search_result header file to incorporate
a search through the PWAS table for the customid in addition to the other
criteria normally used for search.  This adds an or'd statement of that
same search instead of modifying the original or specifically replacing
it.  The "replacement" likely would be faster, but maintenance of the
specific query would be required to ensure that the "format" is consistent
with Zen Cart versions, adding to the existing query provides easier
code maintenance because only the specific query associated with the
customid or whatever other data needs to be modified and the base
search code continues to function.