Finalize appended query of product by cusotmid *and* other info.
This query has been written to supplement an existing query (default query)
but to also add the ability to search product on the customid in addition
to possibly limiting the search through the description by category_id if
considered necessary. The limitation though will be of no value if
this query truly is `OR`ed with the existing query as it too already
searches the description.  It is a simple matter to make this the query
for the site; however, seems that it should be able to add to that
query rather than replace it.