Remove class ReflectionFunction from session (#71)
Because SBA stores data as a session, it appears that in `PHP 7.4.3` that
it is unable to properly address the `ReflectionFunction` class when it
was used the way it was in this module.  To continue use of its functionality
the result has been refactored to support `PHP 7.4.3`.

The message received (with particular data "removed") was:
````php
[19-May-2020 14:24:19 Location] Request URI: /favicon.ico, IP address: IPADDRESS
#1  session_start() called at [includes/functions/sessions.php:117]
#2  zen_session_start() called at [includes/init_includes/init_sessions.php:89]
#3  require(includes/init_includes/init_sessions.php) called at [includes/autoload_func.php:27]
#4  require(includes/autoload_func.php) called at [includes/application_top.php:220]
#5  require(includes/application_top.php) called at [index.php:26]
--> PHP Warning: Erroneous data format for unserializing 'ReflectionFunction' in includes/functions/sessions.php on line 117.

[19-May-2020 14:24:19 Location] Request URI: /favicon.ico, IP address: IPADDRESS
#1  session_start() called at [includes/functions/sessions.php:117]
#2  zen_session_start() called at [includes/init_includes/init_sessions.php:89]
#3  require(includes/init_includes/init_sessions.php) called at [includes/autoload_func.php:27]
#4  require(includes/autoload_func.php) called at [includes/application_top.php:220]
#5  require(includes/application_top.php) called at [index.php:26]
--> PHP Warning: session_start(): Failed to decode session object. Session has been destroyed in includes/functions/sessions.php on line 117.
````
Line numbers may not align exactly with a vanilla install as a result of some development code (trigger_error type messages and/or included comments) being present; however, that information has been found not to be a cause of this issue.