Further refines the customid display position
This adds a few options to where to locate the customid in relation to
known items being built for display.
There no longer is just a `'left'`, `'right'`, and `''`.  This was in
part resulting in the text added by SBA to be either on the left or the
right of ZC default text as a unit.  The intention was to allow some
positioning to be offered for the customid with available quantiity to
be placed in one relative location (immediately after the default ZC
products options text.

Now the options are:
`'leftall'` - Place on the far left of the default text.
`'leftstock'` - Place to the left of the stock text.
`'rightstock'` - Place just to the right of the stock text.
`'rightall'` - Place after all other text has been generated.
`''` - Override customid display and don't display or if code is
  added, do whatever is assigned for an empty value.

Currently the difference between `'rightstock'` and `'rightall'` is that
if an attribute has an effect on price, then the attribute's price is
shown after if `'rightstock'` or before if `'rightall'`, with the
alternate price value otherwise shown at the end of the text.