Eliminate PHP notice in checking variable value. (#77)
This tests to see if `$new_price` is set, if it is set, then use that
value (whether it is zero or not).  I still am not sure exactly why it
is in here, but may exist for some other use.  In absence of it
being set (default) use the value on the right.  It offers a sort of
test variable; however, such testing could have been accomplished by
other means.