Refactor and improve NOTIFY_ATTRIBUTES_MODULE_DEFAULT_SWITCH (#81)
Modified the logic to more clearly identify which combinations of options
appear to support always allow the attribute to remain active. There may
be a reason downstream to deactivate the option, but these factors are
not definite disable options.

Added some code comments to explain why the option is used and where
necessary some potential code pitfalls.