Incorporate includes/modules/attributes.php where needed. (#93)
In ZC 1.5.7, changes were made to includes/modules/attributes.php to where
SBA no longer needed to modify the file.  This means that the file does
not need to be included in the plugin distribution for ZC 1.5.7 and above.

This moves the file into the applicable ZC versions and removes it from
the previous location.