attributes_controller prevent SBA interaction if not installed
Prevents causing changes to the attributes_controller normal display
if SBA is not installed to include missing table definition and if
defined if the table is not loaded to the database.

This has only been done for ZC 1.5.6 and 1.5.7.

Normally the expectation would be that an interface would not be updated
if the core of the software wasn't installed.  That said, an improved
means to incorporate this software is to use some other outside merge
such as javascript or jQuery on page load.

As part of this change, also updated the display associated with
whether the product's attributes were being tracked or not. Code had been
consolidated to minimize the majority of content; however, the actual
status had not been updated to reflect that it was/was not tracked
by SBA.