Address `zen_not_null` and upcoming changes
The Zen Cart function `zen_not_null` is changing as a result of PHP 8.x,
the desire to implement its operation to the way it was intended and
that it had been incorrectly used here in this plugin.  Based on an evaluation
of use of the function `zen_not_null` changes were seen as necessary. In
some cases the change was to ensure that it is/was possible to allow a
variable to be set as 'NULL' or 'null' for possibly older systems.